% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_xgboost.R
\name{pred_xgboost_vector}
\alias{pred_xgboost_vector}
\title{Classification from vector of features}
\usage{
pred_xgboost_vector(input = c(0, 0, 0, 0))
}
\description{
Classification from vector of features
}
